# Admin access JS performance improvements