--- a/app/assets/javascripts/common/admins.js
+++ b/app/assets/javascripts/common/admins.js
@@ -231,29 +231,20 @@ AdminAccessInvite.prototype = Object.assign(AdminAccessInvite.prototype, {
     this.selectAllFacilitiesInput().checked = checkboxes.every(checkbox => checkbox.checked)
   },
 
+  parentID: function(checkbox) {
+    return checkbox.dataset.parentId
+  },
+
   updateIndeterminateCheckboxes: function () {
+    let _self = this
+
     // list of all checkboxes under facilityAccess
     const checkboxes = nodeListToArray(ACCESS_LIST_INPUT_SELECTOR, this.facilityAccess)
+    const oneChildPerParent = distinctByFn(checkboxes, this.parentID);
 
-    function groupByParentId(objectArray) {
-      return objectArray.reduce(function (acc, obj) {
-        let key = obj.dataset.parentId
-        if (!acc[key]) {
-          acc[key] = []
-        }
-        acc[key].push(obj)
-        return acc
-      }, {})
-    }
-
-    let elementsByParentId = groupByParentId([...$('[data-parent-id]')]);
-
-    // go through all the checkboxes that are pre-checked and update their parents accordingly
-    for (const parentId in elementsByParentId) {
-      let _self = this
-
+    for (const checkbox of oneChildPerParent) {
       requestAnimationFrame(function () {
-        _self.updateParentCheckedState(elementsByParentId[parentId][0], ACCESS_LIST_INPUT_SELECTOR)
+        _self.updateParentCheckedState(checkbox, ACCESS_LIST_INPUT_SELECTOR)
       })
     }
 
@@ -338,3 +329,11 @@ const nodeListToArray = (selector, parent = document) =>
 // return a function that checks if element contains class
 const containsClass = (className) => ({classList}) =>
   classList && classList.contains(className)
+
+const distinctByFn = function (collection, fn) {
+  return Object.values(
+      collection.reduce(function (accumulator, item) {
+        accumulator[fn(item)] = item
+        return accumulator
+      }, {}))
+}