**Story card:** [ch4550](https://app.clubhouse.io/simpledotorg/story/4550/admin-access-page-is-timing-out)

## Because

Follow up to #2848. The admin access page is very slow to load with a large number of facilities. On IHCI it takes roughly ~110 seconds to render the page.

## This addresses

The load time for first paint (~14k facilities) is roughly:

- 8 seconds - AJAX and server side rendering of the access tree markup from `/admins/<user_id>/access_tree/edit`
- < 2 seconds - downloading the response (size of the tree's markup is ~800kb)
- < 1 second - parsing the HTML
- ~90 seconds - JS time, painting the tree

This optimises a loop in JS which updates the parent checkbox's status based on the children's checkboxes.The following was being done recursively for each checkbox at the first paint:
- Find parent of each checkbox
- Find all children of the parent, see if all, some or none are checked
- Update the parent checkbox
- Do this for parent of current checkbox

This avoids calling the loop on all checkboxes. Instead we call it only once per parent. This cuts down the JS time to 3-4 seconds.

Before (with 5k facilities):
![image](https://user-images.githubusercontent.com/16774200/130068918-5009bada-1724-4fe2-b8eb-22ec6b9ab36a.png)

After (with 5k facilities): 
![image](https://user-images.githubusercontent.com/16774200/130068989-2c0c7a49-bc48-421c-8c84-01eb227a1b07.png)