@rsanheim 2021-08-19
Is this a follow up to https://github.com/simpledotorg/simple-server/pull/2848 ?

Please do add cross refence PR's to your description(s) for PRs like this.  It makes it much easier to track fixes that are broken down along several PRs for a single story or bug fix.  It makes it possible to see much more of the big picture via the PR cross references.