@prabhanshuguptagit 2021-08-20
> Why is this part so low?

This takes the most amount of time _now_, actually. The JS is fairly scalable now, and it scales linearly with the number of facility groups (not facilities), so we can consider it a non-problem for the medium-term. We are mostly bottlenecked by this time.

I'll be catching up with claudio to chalk out what fetching the tree asynchronously would look like.